refactor : 수익률을 구하는 기능을 LottoResult 클래스가 맡아서 하도록 수정
